[AIFlow][Refactor] Move method from meta/__init__.py to client/__init__.py
